[SlateEditor] apply suggestions
